Implement `SignatureProvider` for `Ecdsa`
